requirements.txt
youtube-dl==2021.6.6
discord==1.0.1
discord.py==1.7.3
pandas==1.2.4
pandas-datareader==0.9.0rc1
python-dotenv==0.18.0
requests~=2.25.1